Changing function ptr typedef Socket-> t_client * -> ID id

+++ server/src/core/senders.c
@@ -17,7 +17,7 @@ int send_to_gui(t_server *server, char const *fmt, ...)
   if (vsnprintf(cmd, MESSAGE_MAX_SIZE, fmt, va) < 0)
     return (1);
   va_end(va);
-  strncircular(&server->gui.w, cmd, strlen(cmd));
+  strncircular(&server->gui.w, cmd, (int)strlen(cmd));
   return (0);
 }
 
@@ -31,6 +31,6 @@ int send_to_ia(t_server *server, Socket sock, char const *fmt, ...)
   if (vsnprintf(cmd, MESSAGE_MAX_SIZE, fmt, va) < 0)
     return (1);
   va_end(va);
-  strncircular(&server->game.clients[sock].w, cmd, strlen(cmd));
+  strncircular(&server->game.clients[sock].w, cmd, (int)strlen(cmd));
   return (0);
 }
